Update Core Code
This is the code portion of the FSE. This is a version with all files merged together (though this harms performance, it was a lot more simple to submit and for this project the purpose was to pass, not to exactly organize). The FLA file that accompanies it is not able to be uploaded due to size. For this reason, it can be shared with those who care enough to see by contacting me via email.